Add ProtectedRoute component and lint setup
